Fix particle types

+++ src/lingon_sylt.rs
@@ -355,7 +355,7 @@ macro_rules! particle_prop {
             "sylt::lingon_sylt"
             $name
             "Sets the given particle prop"
-            [Two(String(name), Int(system)), Two(Int(lo), Int(hi))] -> Type::Void => {
+            [Two(String(name), Int(system)), Two(Float(lo), Float(hi))] -> Type::Void => {
                 if name.as_ref() != "particle" {
                     return error!("l_gfx_particle_spawn", "Expected a particle system ID");
                 }
@@ -365,7 +365,7 @@ macro_rules! particle_prop {
                 });
                 Ok(Nil)
             },
-            [Two(String(name), Int(system)), Two(Int(lo), Int(hi)), One(String(dist))] -> Type::Void => {
+            [Two(String(name), Int(system)), Two(Float(lo), Float(hi)), One(String(dist))] -> Type::Void => {
                 if name.as_ref() != "particle" {
                     return error!("l_gfx_particle_spawn", "Expected a particle system ID");
                 }
@@ -418,7 +418,7 @@ sylt_macro::extern_function!(
     "sylt::lingon_sylt"
     l_gfx_particle_start_color
     "Sets the spawn color of the particles"
-    [Two(String(name), Int(system)), Three(Int(r), Int(g), Int(b))] -> Type::Void => {
+    [Two(String(name), Int(system)), Three(Float(r), Float(g), Float(b))] -> Type::Void => {
         if name.as_ref() != "particle" {
             return error!("l_gfx_particle_spawn", "Expected a particle system ID");
         }
@@ -433,7 +433,7 @@ sylt_macro::extern_function!(
         });
         Ok(Nil)
     },
-    [Two(String(name), Int(system)), Four(Int(r), Int(g), Int(b), Int(a))] -> Type::Void => {
+    [Two(String(name), Int(system)), Four(Float(r), Float(g), Float(b), Float(a))] -> Type::Void => {
         if name.as_ref() != "particle" {
             return error!("l_gfx_particle_spawn", "Expected a particle system ID");
         }
@@ -456,7 +456,7 @@ sylt_macro::extern_function!(
     "sylt::lingon_sylt"
     l_gfx_particle_end_color
     "Sets the spawn color of the particles"
-    [Two(String(name), Int(system)), Three(Int(r), Int(g), Int(b))] -> Type::Void => {
+    [Two(String(name), Int(system)), Three(Float(r), Float(g), Float(b))] -> Type::Void => {
         if name.as_ref() != "particle" {
             return error!("l_gfx_particle_spawn", "Expected a particle system ID");
         }
@@ -471,7 +471,7 @@ sylt_macro::extern_function!(
         });
         Ok(Nil)
     },
-    [Two(String(name), Int(system)), Four(Int(r), Int(g), Int(b), Int(a))] -> Type::Void => {
+    [Two(String(name), Int(system)), Four(Float(r), Float(g), Float(b), Float(a))] -> Type::Void => {
         if name.as_ref() != "particle" {
             return error!("l_gfx_particle_spawn", "Expected a particle system ID");
         }
